Added logic to button

+++ app/src/main/java/com/example/passfort/repository/PasswordsDetailRepo.kt
@@ -3,7 +3,7 @@ package com.example.passfort.repository
 import com.example.passfort.model.dbentity.PasswordRecordEntity
 
 interface PasswordsDetailRepo {
-    suspend fun getPassword(id: Int): PasswordRecordEntity
+    suspend fun getPassword(id: Long): PasswordRecordEntity
     suspend fun upsertPassword(password: PasswordRecordEntity)
     suspend fun deletePassword(password: PasswordRecordEntity)
 }
